ref(processor): Split tests for minidumps and resolution
