+++ tests/minidump.rs
@@ -5,7 +5,7 @@ mod common;
 
 use std::collections::BTreeSet;
 
-use breakpad::{CodeModule, ProcessState, Resolver};
+use breakpad::ProcessState;
 use common::{assert_snapshot, fixture_path};
 
 /// Process a minidump file
@@ -22,28 +22,7 @@ fn get_minidump_process_state() {
 #[test]
 fn obtain_referenced_modules() {
     let state = process_minidump("electron.dmp");
-    let modules: BTreeSet<&CodeModule> =
-        state.referenced_modules().iter().cloned().collect();
+    let modules: BTreeSet<_> = state.referenced_modules().iter().cloned().collect();
 
     assert_snapshot("referenced_modules.txt", &modules);
-}
-
-#[test]
-fn resolve_electron_stack_frame() {
-    let state = process_minidump("electron.dmp");
-    let thread = state.threads().first().unwrap();
-    let frame = thread.frames()[1];
-
-    let resolver = Resolver::new(fixture_path("Electron Framework.sym"))
-        .expect("Could not load symbols for Electron Framework.");
-
-    let resolved_frame = resolver.resolve_frame(&frame);
-    assert_snapshot("resolved_frame.txt", &resolved_frame);
-}
-
-#[test]
-fn create_corrupt_resolver() {
-    let resolver = Resolver::new(fixture_path("Corrupt.sym"))
-        .expect("Could not load symbols for Corrupt.");
-    assert!(resolver.corrupt());
 }
